chore: update namespaces for relocated EF configurations
Follow-up to a344f5c97, which moved the three files but left the
namespace/using edits uncommitted. Adds the actual:

- AccountMergeRequestConfiguration: namespace .Profiles -> .Users
- EventParticipationConfiguration: namespace .Shifts -> .Users
- GoogleResourceConfiguration: namespace Configurations -> Configurations.GoogleIntegration
  + [Grandfathered("HUM0024", ...)] (see PR description)

Fixes nobodies-collective/Humans#981
