+++ src/Humans.Infrastructure/Data/Configurations/GoogleIntegration/GoogleResourceConfiguration.cs
@@ -1,9 +1,15 @@
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 using Humans.Domain.Entities;
+using Humans.Application.Architecture;
 
-namespace Humans.Infrastructure.Data.Configurations;
+namespace Humans.Infrastructure.Data.Configurations.GoogleIntegration;
 
+[Grandfathered(
+    ruleId: "HUM0024",
+    justification: "Pre-existing cross-section EF navigation join; migrating to bare FK + service-level stitching.",
+    since: "2026-05-25",
+    issueRef: "docs/architecture/roslyn-analysis.md#hum0024")]
 public class GoogleResourceConfiguration : IEntityTypeConfiguration<GoogleResource>
 {
     public void Configure(EntityTypeBuilder<GoogleResource> builder)
